persist: split state encoding out into a new module
